requirements.txt
selenium==3.141.0
requests==2.25.1
pillow==9.0.1
python-slugify==6.1.2